Don't submit more than 25 WiFi networks when calling the Google Maps API.
When there are too many WiFi networks the constructed Google Maps API
request becomes too large, and Google returns an error. This patch
limits the request to 25 networks, and also adds error handling so the
report is still sent even if Google returns an error.

This fixes prey-bash-client issues #395 and #396.